+++ packages/carbon-graphs/src/js/controls/Scatter/Scatter.js
@@ -236,6 +236,22 @@ class Scatter extends GraphContent {
         this.dataTarget,
         graph.config,
       );
+    } else if (graph.config.axis.y2.show) {
+      /* If data belongs to yAxis and shapes present in yAxis is null then prepareLabelShapeItem method gets executed to add shapes in yAxis
+       else if data belongs to y2Axis and shapes present in yAxis is null then prepareLabelShapeItem method gets executed to add shapes in y2Axes. */
+      if (this.config.yAxis === 'y' && (document.querySelector('.carbon-y-axis-label-shape-container svg') == null)) {
+        prepareLabelShapeItem(
+          graph.config,
+          this.dataTarget,
+          graph.axesLabelShapeGroup[this.config.yAxis],
+        );
+      } else if (this.config.yAxis === 'y2' && (document.querySelector('.carbon-y2-axis-label-shape-container svg') == null)) {
+        prepareLabelShapeItem(
+          graph.config,
+          this.dataTarget,
+          graph.axesLabelShapeGroup[this.config.yAxis],
+        );
+      }
     }
     this.valuesRange = calculateValuesRange(
       this.config.values,